backend/requirements.txt
asgiref==3.7.2
boto3==1.28.56
botocore==1.31.56
Django==4.2.5
django-cors-headers==4.2.0
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
